+++ cpp/core/multivariate_gauss.cpp
@@ -10,20 +10,20 @@ using namespace std;
 
 //output: sample set
 
-VectorXf multivariate_gauss(VectorXf x, MatrixXf P, int n) 
+VectorXd multivariate_gauss(VectorXd x, MatrixXd P, int n) 
 {
 	int len = x.size();
 	//choleksy decomposition
-	MatrixXf S = P.llt().matrixL();
-	MatrixXf X(len,n);
+	MatrixXd S = P.llt().matrixL();
+	MatrixXd X(len,n);
 	
 
-    float LO = -1.0f;
-    float HI = 1.0f;
+    double LO = -1.0f;
+    double HI = 1.0f;
 
     for (int i = 0; i < len; i++) {
         for (int j=0; j< n; j++) {
-            float r3 = LO + (float)rand()/((float)RAND_MAX/(HI-LO));
+            double r3 = LO + (double)rand()/((double)RAND_MAX/(HI-LO));
             X(i,j) = r3;
         }
     }
@@ -31,7 +31,7 @@ VectorXf multivariate_gauss(VectorXf x, MatrixXf P, int n)
     //TODO: this does not work. Also fixed other instances of nRandMat	
 	//X = nRandMat::randn(len,n);	
 	
-	MatrixXf ones = MatrixXf::Ones(1,n);	
+	MatrixXd ones = MatrixXd::Ones(1,n);	
 	return S*X + x*ones;
 }
