+++ cpp/fastslam2/predict.cpp
@@ -8,10 +8,10 @@
 
 using namespace std;
 
-void predict(Particle &particle,float V,float G,Matrix2f Q, float WB,float dt, int addrandom)
+void predict(Particle &particle,double V,double G,Matrix2d Q, double WB,double dt, int addrandom)
 {
-	VectorXf xv = particle.xv();
-        MatrixXf Pv = particle.Pv();
+	VectorXd xv = particle.xv();
+        MatrixXd Pv = particle.Pv();
         #if 0
         cout<<"particle.xv() in predict"<<endl;
         cout<<particle.xv()<<endl;
@@ -19,19 +19,19 @@ void predict(Particle &particle,float V,float G,Matrix2f Q, float WB,float dt, i
         cout<<particle.Pv()<<endl;
 	#endif
         //Jacobians
-	float phi = xv(2);
-	MatrixXf Gv(3,3); 
+	double phi = xv(2);
+	MatrixXd Gv(3,3); 
 
 	Gv << 1,0,-V*dt*sin(G+phi),
 	       0,1,V*dt*cos(G+phi),
 	       0,0,1;
-	MatrixXf Gu(3,2); 
+	MatrixXd Gu(3,2); 
 	Gu << dt*cos(G+phi), -V*dt*sin(G+phi),
 		dt*sin(G+phi), V*dt*cos(G+phi),
 		dt*sin(G)/WB, V*dt*cos(G)/WB;
 
 	//predict covariance
-	Matrix3f newPv;//(Pv.rows(),Pv.cols());
+	Matrix3d newPv;//(Pv.rows(),Pv.cols());
         
         //TODO: Pv here is somehow corrupted. Probably in sample_proposal
         #if 0
@@ -44,24 +44,24 @@ void predict(Particle &particle,float V,float G,Matrix2f Q, float WB,float dt, i
 
 	//optional: add random noise to predicted state
 	if (addrandom ==1) {
-		VectorXf A(2);
+		VectorXd A(2);
 		A(0) = V;
 		A(1) = G;
-		VectorXf VG(2);
+		VectorXd VG(2);
 		VG = multivariate_gauss(A,Q,1);	
 		V = VG(0);
 		G = VG(1);	
 	}	
 
 	//predict state
-	Vector3f xv_temp(3);
+	Vector3d xv_temp(3);
         xv_temp << xv(0) + V*dt*cos(G+xv(2)),
 	           xv(1) + V*dt*sin(G+xv(2)),
 	           pi_to_pi2(xv(2) + V*dt*sin(G/WB));
         particle.setXv(xv_temp);
 }
 
-float pi_to_pi2(float ang) 
+double pi_to_pi2(double ang) 
 {
     if (ang > pi) {
         ang = ang - (2*pi);
